Continue renaming drawable -> widget
