+++ eco_perf/terminal_interface/graphics/terminal.h
